Renamed Draft Release Workflow
Renamed the release.yml to draft-release.yml since it creates a draft release and not a published one. Also, the workflow changed to tag the pushed image as draft and not as latest. The release.yml workflow added which runs post-release in order to tag the draft docker image as latest.

Signed-off-by: tdakanalis <tdakanalis@cytech.gr>