feat(backend): Add `--user-session` CLI Parameter
Adds the --user-session CLI parameter so that users can pass the
user_session cookie. The GitHub personal token works only for
authenticating users to download images but not for any other file type.
As a workaround, the tool will send the user_session cookie when
migrating assets and files from a private GitHub repository.